requirements.txt
numpy
pandas
websocket-client
eth-account>=0.10
hyperliquid-python-sdk